Split CopyOnWriteSkipListStateMapTest into basic and complex sets.
The complex set mainly includes snapshot related cases, such as copy-on-write contract,
logical/physical remove correctness and snapshot prune useless versions, etc.

Have also confirmed no test coverage reduction, CopyOnWriteSkipListStateMap 99% (546/551)
and CopyOnWriteSkipListStateMap 98% (77/78) line coverage.
